+++ dolphinscheduler-common/pom.xml
@@ -588,11 +588,6 @@
         </dependency>
 
 		<dependency>
-			<groupId>org.springframework</groupId>
-			<artifactId>spring-context</artifactId>
-		</dependency>
-
-		<dependency>
 			<groupId>org.codehaus.janino</groupId>
 			<artifactId>janino</artifactId>
 			<version>${codehaus.janino.version}</version>
